fix(metrics): keep conversions fractional and fix google zero
The conversions metric was typed Int64 end to end, and the Google Ads and
Microsoft adapters rounded to reach it. Both vendors type their conversion
metric as a double and credit fractional conversions under data-driven,
position-based and offline attribution, so a campaign credited 0.4 of a
conversion was reported as having produced ZERO -- and the no_conversions
rule reads exactly that number, so the rounding MANUFACTURED the finding
the rule exists to report. The field is now Float64 through the design,
domain, rules and response, and no adapter rounds.

Microsoft rounded per report ROW before summing, so twenty rows of 0.4
totalled 0 for a campaign that converted eight times; it now accumulates
the float. LinkedIn's externalWebsiteConversions is a long and carries no
fraction, so its exact int64 overflow guard is kept and only widened at
the boundary.

Separately, an absent metrics.conversions from Google Ads is a MEASURED
zero, not unmeasured. The field is always in the SELECT list and Google
Ads REST is proto3 JSON, which omits default-valued fields -- the same
reason parseMetricInt already treats an omitted impressions value as 0.
Leaving it nil meant no_conversions could never fire for a Google
campaign that genuinely converted nobody.

nil keeps its meaning for the four platforms that cannot report a
campaign-level conversion count at all (Meta, X, Reddit, email): the
field stays absent on the wire and the rule never fires. A zero default
there would be the fabricated measurement this epic removes.

A mutant survived the first pass: loosening the rule's comparison to
'< 1' broke no test, because every fixture used a whole number. That is
the defect this change exists to prevent, invisible to the suite meant to
cover it; a fractional case now kills it.

+++ internal/platform/linkedin/metrics.go
@@ -269,14 +269,18 @@ func (c *Client) GetCampaignMetrics(ctx context.Context, accountID, campaignID s
 			if conv < 0 {
 				return nil, fmt.Errorf("get campaign metrics: negative externalWebsiteConversions in response element")
 			}
+			// The running total is accumulated as an int64 and only widened at the end.
+			// externalWebsiteConversions is typed `long` in LinkedIn's Ads Reporting schema —
+			// unlike Google's and Microsoft's doubles, it carries no fraction — so the exact
+			// integer overflow guard is kept rather than being lost to float arithmetic.
 			var running int64
 			if metrics.Conversions != nil {
-				running = *metrics.Conversions
+				running = int64(*metrics.Conversions)
 			}
 			if conv > math.MaxInt64-running {
 				return nil, fmt.Errorf("get campaign metrics: aggregate externalWebsiteConversions overflows int64")
 			}
-			total := running + conv
+			total := float64(running + conv)
 			metrics.Conversions = &total
 		}
 		if elem.CostInUsd != nil {
